fix(mobile): resolve iOS SwiftLint and Android compile CI failures
- Android: fix ConferenceLogic.icalUrl return type (URL -> String)
- Android: add Robolectric to ConferenceLogicTest for context-dependent tests
- iOS: extract PlannerSnapshotCoding to satisfy PlannerLogic type_body_length
- iOS: split PlannerModel.fetchPlannerSnapshot into helper loaders for function_body_length
