+++ clients/ios/Lextures/Features/Planner/PlannerModel.swift
@@ -82,7 +82,7 @@ final class PlannerModel {
                     accessToken: accessToken
                 )
             }
-            let decoded = PlannerLogic.decodeSnapshot(snapshotResult.value)
+            let decoded = PlannerSnapshotCoding.decodeSnapshot(snapshotResult.value)
             todos = decoded.todos
             events = decoded.events
             if let cached = snapshotResult.cached, cached.isStale(isOnline: NetworkMonitor.shared.isOnline) {
@@ -104,7 +104,48 @@ final class PlannerModel {
         accessToken: String
     ) async throws -> PlannerSnapshot {
         async let notebookTasksTask = LMSAPI.fetchNotebookTasks(accessToken: accessToken)
+        let (structures, grades) = await loadStructuresAndGrades(
+            studentCourses: studentCourses,
+            accessToken: accessToken
+        )
+        let notebookTasks = try await notebookTasksTask
+        let academic = await loadAcademicEvents(studentCourses: studentCourses, accessToken: accessToken)
+        let officeHoursByCourse = await loadOfficeHoursByCourse(
+            studentCourses: studentCourses,
+            accessToken: accessToken
+        )
+        let liveMeetingsByCourse = await loadLiveMeetingsByCourse(
+            studentCourses: studentCourses,
+            accessToken: accessToken
+        )
+        let evaluationStatusByCourse = await loadEvaluationStatuses(
+            studentCourses: studentCourses,
+            accessToken: accessToken
+        )
+        let parentConferenceBookings = await loadParentConferenceBookingsIfNeeded(accessToken: accessToken)
+        let todos = PlannerLogic.collectTodos(
+            studentCourses: studentCourses,
+            structureByCourseCode: structures,
+            notebookTasks: notebookTasks,
+            gradesByCourseCode: grades,
+            evaluationStatusByCourseCode: evaluationStatusByCourse
+        )
+        let events = PlannerLogic.collectCalendarEvents(
+            studentCourses: studentCourses,
+            structureByCourseCode: structures,
+            notebookTasks: notebookTasks,
+            academicEvents: academic,
+            officeHoursByCourseCode: officeHoursByCourse,
+            liveMeetingsByCourseCode: liveMeetingsByCourse,
+            parentConferenceBookings: parentConferenceBookings
+        )
+        return PlannerSnapshotCoding.encodeSnapshot(todos: todos, events: events)
+    }
 
+    private static func loadStructuresAndGrades(
+        studentCourses: [CourseSummary],
+        accessToken: String
+    ) async -> (structures: [String: [CourseStructureItem]], grades: [String: MyGradesResponse]) {
         var structures: [String: [CourseStructureItem]] = [:]
         var grades: [String: MyGradesResponse] = [:]
         await withTaskGroup(of: (String, [CourseStructureItem]?, MyGradesResponse?).self) { group in
@@ -126,13 +167,14 @@ final class PlannerModel {
                 if let myGrades { grades[code] = myGrades }
             }
         }
+        return (structures, grades)
+    }
 
-        let notebookTasks = try await notebookTasksTask
+    private static func loadAcademicEvents(
+        studentCourses: [CourseSummary],
+        accessToken: String
+    ) async -> [AcademicCalendarEvent] {
         var academic: [AcademicCalendarEvent] = []
-        var officeHoursByCourse: [String: OfficeHoursAvailability] = [:]
-        var liveMeetingsByCourse: [String: [VirtualMeeting]] = [:]
-        var parentConferenceBookings: [ParentConferenceBooking] = []
-        var evaluationStatusByCourse: [String: EvaluationStatus] = [:]
         var seenKeys = Set<String>()
         for course in studentCourses {
             guard let orgId = course.orgId else { continue }
@@ -145,7 +187,14 @@ final class PlannerModel {
             )) ?? []
             academic.append(contentsOf: rows)
         }
+        return academic
+    }
 
+    private static func loadOfficeHoursByCourse(
+        studentCourses: [CourseSummary],
+        accessToken: String
+    ) async -> [String: OfficeHoursAvailability] {
+        var officeHoursByCourse: [String: OfficeHoursAvailability] = [:]
         await withTaskGroup(of: (String, OfficeHoursAvailability?).self) { group in
             for course in studentCourses where course.isOfficeHoursEnabled {
                 group.addTask {
@@ -160,7 +209,14 @@ final class PlannerModel {
                 if let availability { officeHoursByCourse[code] = availability }
             }
         }
+        return officeHoursByCourse
+    }
 
+    private static func loadLiveMeetingsByCourse(
+        studentCourses: [CourseSummary],
+        accessToken: String
+    ) async -> [String: [VirtualMeeting]] {
+        var liveMeetingsByCourse: [String: [VirtualMeeting]] = [:]
         await withTaskGroup(of: (String, [VirtualMeeting]?).self) { group in
             for course in studentCourses where course.isLiveSessionsEnabled {
                 group.addTask {
@@ -175,7 +231,14 @@ final class PlannerModel {
                 if let meetings { liveMeetingsByCourse[code] = meetings }
             }
         }
+        return liveMeetingsByCourse
+    }
 
+    private static func loadEvaluationStatuses(
+        studentCourses: [CourseSummary],
+        accessToken: String
+    ) async -> [String: EvaluationStatus] {
+        var evaluationStatusByCourse: [String: EvaluationStatus] = [:]
         await withTaskGroup(of: (String, EvaluationStatus?).self) { group in
             for course in studentCourses {
                 group.addTask {
@@ -192,36 +255,26 @@ final class PlannerModel {
                 }
             }
         }
+        return evaluationStatusByCourse
+    }
 
-        if let children = try? await LMSAPI.fetchParentChildren(accessToken: accessToken), !children.isEmpty {
-            let childTuples = children.map { child in
-                let name = child.displayName?.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty == false
-                    ? child.displayName!
-                    : child.email
-                return (studentId: child.studentUserId, childName: name)
-            }
-            parentConferenceBookings = await ConferenceLogic.loadParentBookings(
-                children: childTuples,
-                accessToken: accessToken
-            )
+    private static func loadParentConferenceBookingsIfNeeded(accessToken: String) async -> [ParentConferenceBooking] {
+        guard let children = try? await LMSAPI.fetchParentChildren(accessToken: accessToken), !children.isEmpty else {
+            return []
         }
+        return await loadParentConferenceBookings(children: children, accessToken: accessToken)
+    }
 
-        let todos = PlannerLogic.collectTodos(
-            studentCourses: studentCourses,
-            structureByCourseCode: structures,
-            notebookTasks: notebookTasks,
-            gradesByCourseCode: grades,
-            evaluationStatusByCourseCode: evaluationStatusByCourse
-        )
-        let events = PlannerLogic.collectCalendarEvents(
-            studentCourses: studentCourses,
-            structureByCourseCode: structures,
-            notebookTasks: notebookTasks,
-            academicEvents: academic,
-            officeHoursByCourseCode: officeHoursByCourse,
-            liveMeetingsByCourseCode: liveMeetingsByCourse,
-            parentConferenceBookings: parentConferenceBookings
-        )
-        return PlannerLogic.encodeSnapshot(todos: todos, events: events)
+    private static func loadParentConferenceBookings(
+        children: [ParentChildSummary],
+        accessToken: String
+    ) async -> [ParentConferenceBooking] {
+        let childTuples = children.map { child in
+            let name = child.displayName?.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty == false
+                ? child.displayName!
+                : child.email
+            return (studentId: child.studentUserId, childName: name)
+        }
+        return await ConferenceLogic.loadParentBookings(children: childTuples, accessToken: accessToken)
     }
 }
